add debug

+++ preprocessing/create_train_data.py
@@ -65,11 +65,11 @@ class DataParser():
 
     def _convert_1d_np_arr_with_possible_padding(self, lst,
             max_dim, pad_value):
-        np_arr = np.array(lst, dtype=np.int32)[:max_dim]
+        np_arr = np.array(lst)[:max_dim]
         return np.pad(np_arr,
                       pad_width=(0, max_dim - np_arr.shape[0]),
                       mode="constant",
-                      constant_values=pad_value)
+                      constant_values=pad_value).astype(np.int32)
 
     def _parse_data_from_tokens_list(self, tokens_list, tokens_ner_dict):
         """Input: A spaCy doc.
